refactor confetti into its own file
